ref: combined layout on (regflow) path to serve on both paths like sign-in and sign-out; layout is now using usePathname hook to determine which text is added to h1 and p of left column
