+++ apps/web/app/(regflow)/sign-up/page.tsx
@@ -1,0 +1,7 @@
+export default function SignInPage() {
+  return (
+    <div>
+      <p>Hello from Sign-Up page</p>
+    </div>
+  );
+}
